feat(seasons): float points, chat announcements, NIC filtering, anti-farming
Schema & types:
- Change total_points and current_value from BIGINT to FLOAT in migration and schema docs
- Update all C# types (long → double) in SeasonModels, SeasonRepository, SeasonService, AdminTool SeasonRepository

SeasonService:
- Add IChannelManager dependency for chat announcements to "Seasons Info" channel
- Post channel messages on season start/end, tier unlock, objective complete, season end
- Update channel topic on season start and end
- Hourly leaderboard announcement (top 10) during active season
- Point calculation changed from floor(long) to round(double, 2)
- Objective progress now incremented by basePoints instead of raw amount

NIC tracking (CharacterWallet):
- Replace sign-based detection with explicit TransactionType switch
- Only record NicSpent/NicEarned for specific meaningful transaction types

PvP kill (Player):
- Skip kill credit if victim is blessed (IsBlessed)
- Skip kill credit if victim and killer share the same account IP (anti-self-farming)

Additional hooks:
- HarvesterModule: mineral mined hook
- TransportAssignment: direct NIC recording for collateral/reward payments

Discord integration (ChannelManager / EventListenerService):
- Forward in-game channel messages to Discord if channel has DiscordId
- Simplify Discord→game message handler (remove unused channel ID filter)

Docs:
- Update seasons design spec: float types, chat announcements section, revised event hooks table, revised mail notifications
- Update admin tool fixes spec: LeaderboardEntryRow type change

+++ docs/Patches/p36.0/Features/Seasons/migration.sql
@@ -54,7 +54,7 @@ CREATE TABLE season_leaderboard_rewards (
 CREATE TABLE season_character_points (
     character_id                 INT      NOT NULL,
     season_id                    INT      NOT NULL REFERENCES seasons(id),
-    total_points                 BIGINT   NOT NULL DEFAULT 0,
+    total_points                 FLOAT           NOT NULL DEFAULT 0,
     last_updated                 DATETIME NOT NULL DEFAULT GETUTCDATE(),
     intro_mail_sent              BIT      NOT NULL DEFAULT 0,
     leaderboard_reward_delivered BIT      NOT NULL DEFAULT 0,
@@ -65,7 +65,7 @@ CREATE TABLE season_objective_progress (
     character_id   INT          NOT NULL,
     season_id      INT          NOT NULL REFERENCES seasons(id),
     objective_id   INT          NOT NULL REFERENCES season_objectives(id),
-    current_value  BIGINT       NOT NULL DEFAULT 0,
+    current_value  FLOAT               NOT NULL DEFAULT 0,
     completed      BIT          NOT NULL DEFAULT 0,
     completed_time DATETIME         NULL,
     bonus_awarded  BIT          NOT NULL DEFAULT 0,
